+++ apps/docs/src/guide/types/snippets/address/utilities-function-1.ts
@@ -1,0 +1,9 @@
+// #region full
+import { Address } from 'fuels';
+
+const address = Address.fromRandom();
+
+const addressCloneFromB256 = Address.fromString(address.toB256());
+// #endregion full
+
+console.log('addressCloneFromB256', addressCloneFromB256);