+++ packages/fuel-gauge/src/predicate/utils/predicate/index.ts
@@ -1,3 +1,3 @@
 export * from './assertBalance';
 export * from './assertBalances';
-export * from './fundPredicate';
+export * from './fundAccount';