+++ packages/fuels/src/cli/commands/dev/autoStartFuelCore.ts
@@ -35,6 +35,7 @@ export const autoStartFuelCore = async (config: FuelsConfig) => {
       loggingEnabled: loggingConfig.isLoggingEnabled,
       basePath: config.basePath,
       fuelCorePath: config.fuelCorePath,
+      includeInitialState: true,
     });
 
     fuelCore = {